smart_meter_analytics/cloud_processing/requirements.txt
apache-beam[gcp]>=2.50.0
google-cloud-pubsub>=2.18.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-monitoring>=2.16.0
python-dateutil>=2.8.2
